+++ dummyapp/db/migrate/20230613195631_create_cases.rb
@@ -1,0 +1,8 @@
+class CreateCases < ActiveRecord::Migration[7.0]
+  def change
+    create_table :cases do |t|
+
+      t.timestamps
+    end
+  end
+end