+++ dummyapp/db/migrate/20230614221012_add_json_text_field_to_cases.rb
@@ -1,0 +1,5 @@
+class AddJsonTextFieldToCases < ActiveRecord::Migration[7.0]
+  def change
+    add_column :cases, :json_text_field, :text
+  end
+end