+++ lib/annotate_rb/model_annotator/single_file_annotator.rb
@@ -26,7 +26,7 @@ module AnnotateRb
           old_content = File.read(file_name)
 
           file_components = FileComponents.new(old_content, annotation, options)
-          builder = FileBuilder.new(file_components, annotation, annotation_position, options)
+          builder = FileBuilder.new(file_components, annotation_position, options)
 
           return false if file_components.has_skip_string?
           return false if !file_components.annotations_changed? && !options[:force]